pyproject.toml
[tool.poetry]
name = "research-roc-at-home"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "ktree", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
numpy = "^1.25.0"
pyyaml = "^6.0.1"
loguru = "^0.7.0"
types-pyyaml = "^6.0.12.11"
pydantic = "^2.3.0"
ipython = "^8.14.0"
networkx = "^3.1"

[[tool.poetry.source]]
name = "rocsys"
url = "https://dl.cloudsmith.io/basic/rocsys/python/python/simple/"
priority = 'explicit'


[tool.poetry.group.lint.dependencies]
flake8 = "^6.1.0"
flake8-isort = "^6.0.0"
flake8-black = "^0.3.6"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
scipy = "^1.11.3"


[tool.poetry.group.debug.dependencies]
debugpy = "^1.6.7.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
# https://github.com/PyCQA/isort/issues/457
no_sections = true

[tool.black]
line-length = 120
preview = true

[tool.mypy]
plugins = [
    "pydantic.mypy",
    "numpy.typing.mypy_plugin",
]
ignore_missing_imports = true
follow_imports = "skip"
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
files = [
    "src",
]

